Cover emit graph library validation
