+++ tests/integration/cli_build.rs
@@ -20,6 +20,8 @@ mod direct_build_graph_validation;
 mod emit_direct;
 #[path = "cli_build/emit_direct_host_effects.rs"]
 mod emit_direct_host_effects;
+#[path = "cli_build/emit_direct_validation.rs"]
+mod emit_direct_validation;
 #[path = "cli_build/frontend_json.rs"]
 mod frontend_json;
 #[path = "cli_build/graph_validation.rs"]
